refactor(#331): [log]/LogConfig を [developer]/DeveloperConfig にリネーム
[log] は「ログレベル/ローテーション付きの汎用ロギング基盤」を匂わせるが、
本 PR はそれを作らない(swift-log 非採用)。診断/デバッグトグルの置き場として
[developer] の方が正直で、将来 dev トグルが増えても素直に同居できる。

- LogConfig -> DeveloperConfig、セクション [log] -> [developer]
- file キーは [developer].file だと曖昧なので lyrics_resolution_file に明示化
  (各トグルが自分のキーを持つ形)
- ARCHITECTURE の Implementation Modules 図に LyricsResolutionLog を追加
  (Support、LyricsRepository からの依存エッジ)
- README/ARCHITECTURE/AGENTS の参照を追随
