+++ AGENTS.md
@@ -96,7 +96,7 @@ Shared conventions:
   metadata sources (LLM/MusicBrainz/Regex) are queried and merged, and every
   lyrics tier (LRCLIB exact match, validated fuzzy search, user
   `fallback_command` script) is tried across all candidates before giving up.
-  An opt-in `[log] lyrics_resolution` trace (#331) records each tier's
+  An opt-in `[developer] lyrics_resolution` trace (#331) records each tier's
   accept/reject with its reason (title similarity, duration delta) to a local
   file for diagnosing intermittent misses; it is off by default and
   behavior-neutral.
