+++ README.md
@@ -464,22 +464,23 @@ own subprocess timeout so one dead endpoint cannot eat the whole budget,
 and remember that `timeout_ms` covers the entire `fallback_command` run —
 budget it above your per-source timeout with headroom.
 
-### `[log]` — lyrics-resolution debug trace (optional)
+### `[developer]` — diagnostic toggles (optional)
 
-Off by default. When a song's lyrics fail to appear (or the wrong lyrics
-show), turn this on to record exactly *why* each attempt was accepted or
+Off by default. A home for opt-in developer/debug toggles — currently the
+lyrics-resolution trace. When a song's lyrics fail to appear (or the wrong
+lyrics show), turn it on to record exactly *why* each attempt was accepted or
 rejected — the generated candidates, and for every tier (LRCLIB exact match,
 validated fuzzy search, custom script) the title-similarity score and
 duration delta behind each accept/reject.
 
 ```toml
-[log]
+[developer]
 lyrics_resolution = true
-# file = "~/my-lyra-trace.log"   # optional; default is below
+# lyrics_resolution_file = "~/my-lyra-trace.log"   # optional; default is below
 ```
 
 - `lyrics_resolution` — enable the trace (default `false`).
-- `file` — where to append. Omit to use
+- `lyrics_resolution_file` — where to append. Omit to use
   `${XDG_CACHE_HOME:-~/.cache}/lyra/lyrics-debug.log`. `~` and
   `$XDG_CACHE_HOME` are honored.
